Make supabase client conditional

+++ src/supabase/profiles.ts
@@ -11,6 +11,7 @@ const LOG = new Logger('profiles');
  * @returns {Promise<Profile | null>} | User data
  */
 export const getProfileById = async (id: string): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.from('profiles').select().eq('id', id).single();
 
@@ -36,6 +37,7 @@ export const updateProfileUsername = async (
     id: string,
     username: string
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.from('profiles')
             .update({ username: username })
@@ -61,6 +63,7 @@ export const updateProfileUsername = async (
  * @returns {Promise<void>}
  */
 export const deleteProfileById = async (id: string): Promise<void> => {
+    if (!SUPABASE) return;
     try {
         // delete profile
         const { data, error } = await SUPABASE.from('profiles').delete().eq('id', id).select();
@@ -92,6 +95,7 @@ export const getProfileArray = async (
     profileId: string,
     whichCol: 'queue' | 'watched' | 'favorites'
 ): Promise<string[] | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.from('profiles')
             .select(whichCol)
@@ -121,6 +125,7 @@ export const addToProfileArray = async (
     showId: string,
     whichCol: 'queue' | 'watched' | 'favorites'
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.rpc('add_item', {
             profile_id: profileId,
@@ -154,6 +159,7 @@ export const removeFromProfileArray = async (
     showId: string,
     whichCol: 'queue' | 'watched' | 'favorites'
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.rpc('remove_item', {
             profile_id: profileId,
@@ -185,6 +191,7 @@ export const clearProfileArray = async (
     profileId: string,
     whichCol: 'queue' | 'watched' | 'favorites'
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.rpc('remove_all', {
             profile_id: profileId,
@@ -215,6 +222,7 @@ export const setProfileAdultFlag = async (
     id: string,
     isAdult: boolean
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.from('profiles')
             .update({ adult: isAdult })
@@ -240,6 +248,7 @@ export const setProfileAdultFlag = async (
  * @returns {Promise<Profile | null>}
  */
 export const setProfileCountry = async (id: string, country: string): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         // TODO: #587 Ensure country code is valid
         const { data, error } = await SUPABASE.from('profiles')
